shuffle some things around;
- consumers include `Toaster` in their app root
- then call `toaster.variant()` wherever
